+++ controller/server/view/network_list_page.ml
@@ -30,9 +30,10 @@ type params =
   { proxy: string option
   ; is_internet_connected: bool
   ; services: Connman.Service.t list
+  ; interfaces: string
   }
 
-let html { proxy; is_internet_connected; services } =
+let html { proxy; is_internet_connected; services; interfaces } =
   Page.html ~current_page:Page.Network (
     div
       [ h1 ~a:[ a_class [ "d-Title" ] ] [ txt "Network" ]
@@ -74,7 +75,14 @@ let html { proxy; is_internet_connected; services } =
           ; if List.length services = 0 then
               txt "No services available"
             else
-              ul ~a:[ a_class [ "d-NetworkList" ] ] (List.map service services)
+              ul
+                ~a:[ a_class [ "d-NetworkList" ]; a_role [ "list" ] ]
+                (List.map service services)
+          ]
+
+      ; section
+          [ h2 ~a:[ a_class [ "d-Subtitle" ] ] [ txt "Network Interfaces" ]
+          ; pre ~a: [ a_class [ "d-Preformatted" ] ]  [ txt interfaces ]
           ]
       ]
   )